src/backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.5.0",
    "dj-database-url>=3.0.1",
    "Django~=5.2.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "google-cloud-bigquery>=3.39.0",
    "gunicorn>=23.0.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pandas-gbq>=0.32.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "skyfield>=1.53",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.9",
]

[tool.ruff]
# 除外するディレクトリ
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "**/migrations/*.py", # マイグレーションファイルはチェックしない
]

# 1行の文字数（Djangoは長くなりがちなので120くらいがおすすめ）
line-length = 120
indent-width = 4

[tool.ruff.lint]
# 有効にするルール
# E: pycodestyle errors
# W: pycodestyle warnings
# F: Pyflakes
# I: isort (importの並び順整理)
# B: flake8-bugbear
# DJ: flake8-django (Django固有のルール)
select = ["E", "W", "F", "I", "B", "DJ"]
ignore = []

# 自動修正可能なルール
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
